Update upload-artifact action and error handling
This commit updates the upload-artifact action from v3 to v4 in the GitHub workflows build file. It also changes the behavior on file reading error from displaying the error message to directly terminating the program execution. Additionally, the notation for including important information in the README file is updated from [!INFO] to [!NOTE].